Refactor graph package:
- Remove legacy files `graph.py` and `graph_agent.py` to streamline the codebase.
- Introduce a new modular structure with dedicated files for agents, checkpointers, decorators, and types.
- Enhance error handling and state management in the `GraphAgent` class.
- Implement an in-memory checkpointer for better state management during execution.
- Add decorators for node execution to standardize behavior across the graph system.
- Introduce a new `StateGraph` engine with improved parallel execution capabilities and monitoring features.

This refactor aims to improve code readability, maintainability, and performance while addressing previous issues related to state corruption and error handling.
