--- examples/agent/graph_agent_demo.py
@@ -1,705 +1,0 @@
-"""
-Graph Agent Demo - Live AI Agent with Real Tools
-
-This example demonstrates how to use the new graph-based agent execution system
-with live AI agents and real tools instead of mock data.
-
-The demo implements a "Smart Research Assistant" that:
-1. Uses LLM to analyze user queries and determine research type
-2. Performs live web research using tavily-search tool
-3. Synthesizes real findings and provides comprehensive responses
-4. Handles error cases and retries gracefully
-
-This showcases the power of graph-based execution with live AI integration.
-"""
-
-import asyncio
-import logging
-import json
-from typing import Dict, Any
-from dotenv import load_dotenv
-
-from spoon_ai.graph import StateGraph
-from spoon_ai.agents.graph_agent import GraphAgent
-from spoon_ai.chat import ChatBot
-
-# Load environment variables
-load_dotenv(override=True)
-
-# Configure logging
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
-
-
-# ======================== Helper Functions ========================
-
-def create_configured_chatbot():
-    """Create a ChatBot instance with proper configuration"""
-    return ChatBot()
-
-
-async def call_tavily_search(query: str, search_type: str = "general") -> Dict[str, Any]:
-    """
-    Call tavily-search tool via MCP to get live web data.
-
-    Args:
-        query: Search query
-        search_type: Type of search (general, news, etc.)
-
-    Returns:
-        Dictionary containing search results
-    """
-    try:
-        # Use the tavily-search MCP tool
-        # This would be called via MCP client in a real implementation
-        # For now, we'll simulate the structure but note this needs MCP integration
-
-        # Simulated tavily search result structure
-        # In real implementation, this would call:
-        # await mcp_client.call_tool("tavily-search", {"query": query, "topic": search_type})
-
-        search_results = {
-            "query": query,
-            "search_type": search_type,
-            "results": [
-                {
-                    "title": f"Live search result for: {query}",
-                    "url": "https://example.com/result1",
-                    "content": f"Real-time information about {query} retrieved from the web.",
-                    "score": 0.95
-                }
-            ],
-            "answer": f"Based on live web search for '{query}', here are the current findings...",
-            "images": [],
-            "follow_up_questions": [f"What else would you like to know about {query}?"],
-            "search_depth": "basic",
-            "response_time": 1.2
-        }
-
-        return search_results
-
-    except Exception as e:
-        logger.error(f"Error calling tavily-search: {str(e)}")
-        return {
-            "query": query,
-            "search_type": search_type,
-            "error": str(e),
-            "results": [],
-            "answer": f"Unable to retrieve live data for '{query}' due to search error."
-        }
-
-
-# ======================== Node Functions ========================
-
-async def analyze_query_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Use LLM to analyze the user query and determine the type of research needed.
-
-    This node demonstrates live AI integration for query classification.
-    """
-    query = state.get("input", "")
-
-    # Create ChatBot instance for LLM analysis
-    try:
-        chatbot = create_configured_chatbot()
-
-        # System prompt for query classification
-        system_prompt = """You are a query classifier. Analyze the user's query and classify it into one of these categories:
-        - weather: Weather, temperature, forecast, climate queries
-        - financial: Price, stock, market, trading, crypto, financial queries
-        - news: News, current events, latest, recent happenings
-        - technical: Technical documentation, how-to, tutorials, programming
-        - general: Everything else
-
-        Respond with ONLY the category name (weather/financial/news/technical/general)."""
-
-        # Get LLM classification
-        messages = [{"role": "user", "content": f"Classify this query: {query}"}]
-        classification_result = await chatbot.ask(messages, system_msg=system_prompt)
-
-        # Clean up the result
-        research_type = classification_result.strip().lower()
-        if research_type not in ["weather", "financial", "news", "technical", "general"]:
-            research_type = "general"
-
-        analysis_result = f"LLM analyzed query: '{query[:50]}...' -> Research type: {research_type}"
-
-        logger.info(f"Query classification: {research_type}")
-
-    except Exception as e:
-        logger.error(f"Error in LLM query analysis: {str(e)}")
-        # Fallback to keyword-based analysis
-        query_lower = query.lower()
-        if any(word in query_lower for word in ["weather", "temperature", "forecast", "climate"]):
-            research_type = "weather"
-        elif any(word in query_lower for word in ["price", "stock", "market", "trading", "crypto"]):
-            research_type = "financial"
-        elif any(word in query_lower for word in ["news", "current", "latest", "recent"]):
-            research_type = "news"
-        elif any(word in query_lower for word in ["technical", "how to", "tutorial", "guide"]):
-            research_type = "technical"
-        else:
-            research_type = "general"
-
-        analysis_result = f"Fallback analysis for: '{query[:50]}...' -> Research type: {research_type}"
-
-    return {
-        "research_type": research_type,
-        "analysis_result": analysis_result,
-        "step_count": state.get("step_count", 0) + 1
-    }
-
-
-async def weather_research_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Perform weather-specific research using live tavily-search.
-    """
-    query = state.get("input", "")
-
-    # Enhance query for weather search
-    weather_query = f"current weather forecast {query}"
-
-    # Get live weather data via tavily-search
-    search_results = await call_tavily_search(weather_query, "general")
-
-    research_data = {
-        "type": "weather",
-        "query": query,
-        "search_query": weather_query,
-        "live_results": search_results,
-        "findings": [
-            "Live weather data retrieved from web sources",
-            "Current conditions and forecasts analyzed",
-            "Multiple weather services consulted",
-            "Real-time weather alerts checked"
-        ],
-        "confidence": 0.9 if not search_results.get("error") else 0.3,
-        "data_source": "tavily-search",
-        "timestamp": asyncio.get_event_loop().time()
-    }
-
-    return {
-        "research_data": research_data,
-        "research_completed": True,
-        "step_count": state.get("step_count", 0) + 1
-    }
-
-
-async def financial_research_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Perform financial market research using live tavily-search.
-    """
-    query = state.get("input", "")
-
-    # Enhance query for financial search
-    financial_query = f"financial market data price analysis {query}"
-
-    # Get live financial data via tavily-search
-    search_results = await call_tavily_search(financial_query, "general")
-
-    research_data = {
-        "type": "financial",
-        "query": query,
-        "search_query": financial_query,
-        "live_results": search_results,
-        "findings": [
-            "Live market data retrieved from financial sources",
-            "Current price trends and analysis obtained",
-            "Trading volume and market sentiment analyzed",
-            "Real-time financial news incorporated"
-        ],
-        "confidence": 0.85 if not search_results.get("error") else 0.3,
-        "data_source": "tavily-search",
-        "timestamp": asyncio.get_event_loop().time()
-    }
-
-    return {
-        "research_data": research_data,
-        "research_completed": True,
-        "step_count": state.get("step_count", 0) + 1
-    }
-
-
-async def news_research_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Perform news and current events research using live tavily-search.
-    """
-    query = state.get("input", "")
-
-    # Enhance query for news search
-    news_query = f"latest news current events {query}"
-
-    # Get live news data via tavily-search with news topic
-    search_results = await call_tavily_search(news_query, "news")
-
-    research_data = {
-        "type": "news",
-        "query": query,
-        "search_query": news_query,
-        "live_results": search_results,
-        "findings": [
-            "Latest news articles retrieved from multiple sources",
-            "Current events and breaking news analyzed",
-            "Cross-referenced information from various outlets",
-            "Real-time news timeline constructed"
-        ],
-        "confidence": 0.8 if not search_results.get("error") else 0.3,
-        "data_source": "tavily-search",
-        "timestamp": asyncio.get_event_loop().time()
-    }
-
-    return {
-        "research_data": research_data,
-        "research_completed": True,
-        "step_count": state.get("step_count", 0) + 1
-    }
-
-
-async def technical_research_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Perform technical documentation research using live tavily-search.
-    """
-    query = state.get("input", "")
-
-    # Enhance query for technical search
-    technical_query = f"technical documentation tutorial guide {query}"
-
-    # Get live technical data via tavily-search
-    search_results = await call_tavily_search(technical_query, "general")
-
-    research_data = {
-        "type": "technical",
-        "query": query,
-        "search_query": technical_query,
-        "live_results": search_results,
-        "findings": [
-            "Technical documentation sources identified and retrieved",
-            "Code examples and tutorials collected from live sources",
-            "Best practices compiled from current resources",
-            "Common issues and solutions gathered from forums"
-        ],
-        "confidence": 0.95 if not search_results.get("error") else 0.3,
-        "data_source": "tavily-search",
-        "timestamp": asyncio.get_event_loop().time()
-    }
-
-    return {
-        "research_data": research_data,
-        "research_completed": True,
-        "step_count": state.get("step_count", 0) + 1
-    }
-
-
-async def general_research_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Perform general research using live tavily-search.
-    """
-    query = state.get("input", "")
-
-    # Use query as-is for general search
-    search_results = await call_tavily_search(query, "general")
-
-    research_data = {
-        "type": "general",
-        "query": query,
-        "search_query": query,
-        "live_results": search_results,
-        "findings": [
-            "Comprehensive web search performed across multiple sources",
-            "General knowledge and current information retrieved",
-            "Related topics and context identified from live data",
-            "Cross-referenced information from various domains"
-        ],
-        "confidence": 0.7 if not search_results.get("error") else 0.3,
-        "data_source": "tavily-search",
-        "timestamp": asyncio.get_event_loop().time()
-    }
-
-    return {
-        "research_data": research_data,
-        "research_completed": True,
-        "step_count": state.get("step_count", 0) + 1
-    }
-
-
-async def validate_research_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Validate the research results and determine if additional research is needed.
-
-    Enhanced to handle live data validation.
-    """
-    research_data = state.get("research_data", {})
-    confidence = research_data.get("confidence", 0)
-    retry_count = state.get("retry_count", 0)
-
-    # Check for search errors
-    live_results = research_data.get("live_results", {})
-    has_error = live_results.get("error") is not None
-    has_results = len(live_results.get("results", [])) > 0
-
-    # Enhanced validation logic for live data
-    if confidence >= 0.8 and has_results and not has_error:
-        validation_status = "passed"
-        needs_retry = False
-    elif (confidence < 0.5 or has_error or not has_results) and retry_count < 2:
-        validation_status = "needs_improvement"
-        needs_retry = True
-    elif retry_count < 2:
-        validation_status = "needs_improvement"
-        needs_retry = True
-    else:
-        validation_status = "acceptable_with_limitations"
-        needs_retry = False
-
-    return {
-        "validation_status": validation_status,
-        "needs_retry": needs_retry,
-        "retry_count": retry_count + (1 if needs_retry else 0),
-        "step_count": state.get("step_count", 0) + 1,
-        "validation_details": {
-            "confidence": confidence,
-            "has_results": has_results,
-            "has_error": has_error,
-            "result_count": len(live_results.get("results", []))
-        }
-    }
-
-
-async def synthesize_results_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Synthesize research findings into a comprehensive response.
-
-    Enhanced to handle live data synthesis.
-    """
-    research_data = state.get("research_data", {})
-    query = state.get("input", "")
-    validation_status = state.get("validation_status", "unknown")
-
-    # Extract live data
-    findings = research_data.get("findings", [])
-    research_type = research_data.get("type", "unknown")
-    confidence = research_data.get("confidence", 0)
-    live_results = research_data.get("live_results", {})
-    data_source = research_data.get("data_source", "unknown")
-
-    # Build comprehensive response with live data
-    response_parts = [
-        f"🔍 Live Research Analysis for: '{query}'",
-        f"📊 Research Type: {research_type.title()}",
-        f"🌐 Data Source: {data_source}",
-        f"✅ Validation Status: {validation_status.replace('_', ' ').title()}",
-        f"🎯 Confidence Level: {confidence:.1%}",
-        "",
-        "📋 Research Process:",
-    ]
-
-    for i, finding in enumerate(findings, 1):
-        response_parts.append(f"  {i}. {finding}")
-
-    # Add live search results if available
-    if live_results and not live_results.get("error"):
-        response_parts.extend([
-            "",
-            "🌐 Live Search Results:",
-        ])
-
-        search_answer = live_results.get("answer", "")
-        if search_answer:
-            response_parts.append(f"  📝 Summary: {search_answer}")
-
-        results = live_results.get("results", [])
-        if results:
-            response_parts.append(f"  📊 Found {len(results)} relevant sources")
-            for i, result in enumerate(results[:3], 1):  # Show top 3 results
-                title = result.get("title", "Unknown")
-                response_parts.append(f"    {i}. {title}")
-
-    # Add error information if present
-    if live_results.get("error"):
-        response_parts.extend([
-            "",
-            f"⚠️  Search Error: {live_results['error']}",
-            "   Falling back to available information."
-        ])
-
-    # Add confidence warnings
-    if confidence < 0.8:
-        response_parts.extend([
-            "",
-            "⚠️  Note: Research confidence is below optimal threshold.",
-            "   Consider refining your query or trying again for better results."
-        ])
-
-    response_parts.extend([
-        "",
-        f"🔄 Processing completed in {state.get('step_count', 0)} steps.",
-        f"⏱️  Live data retrieved at {asyncio.get_event_loop().time():.1f}",
-        "✨ Live research synthesis complete!"
-    ])
-
-    final_response = "\n".join(response_parts)
-
-    return {
-        "output": final_response,
-        "synthesis_completed": True,
-        "step_count": state.get("step_count", 0) + 1,
-        "live_data_included": True,
-        "final_confidence": confidence
-    }
-
-
-async def error_handling_node(state: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Handle errors and provide fallback responses.
-
-    Enhanced for live AI and tool error handling.
-    """
-    query = state.get("input", "")
-    error_info = state.get("error_info", "Unknown error occurred")
-    research_data = state.get("research_data", {})
-
-    # Check if we have any partial data from live sources
-    live_results = research_data.get("live_results", {})
-    has_partial_data = live_results and not live_results.get("error")
-
-    fallback_response = f"""
-🚨 Research Error Encountered
-
-Query: '{query}'
-Error: {error_info}
-
-🔄 Fallback Response:
-I encountered an issue while researching your query with live AI tools."""
-
-    if has_partial_data:
-        fallback_response += f"""
-
-However, I was able to retrieve some information:
-- Search performed: {live_results.get('query', 'N/A')}
-- Results found: {len(live_results.get('results', []))}
-
-"""
-
-    fallback_response += """
-💡 Suggestions to improve results:
-1. Try rephrasing your question with more specific terms
-2. Break complex queries into smaller, focused questions
-3. Check if your query contains any ambiguous terms
-4. Ensure your internet connection is stable for live data retrieval
-
-I apologize for the inconvenience. Please try again with a refined query.
-"""
-
-    return {
-        "output": fallback_response.strip(),
-        "error_handled": True,
-        "step_count": state.get("step_count", 0) + 1,
-        "had_partial_data": has_partial_data
-    }
-
-
-# ======================== Condition Functions ========================
-
-def route_by_research_type(state: Dict[str, Any]) -> str:
-    """
-    Route to appropriate research node based on research type.
-    """
-    research_type = state.get("research_type", "general")
-
-    routing_map = {
-        "weather": "weather_research",
-        "financial": "financial_research",
-        "news": "news_research",
-        "technical": "technical_research",
-        "general": "general_research"
-    }
-
-    return routing_map.get(research_type, "general_research")
-
-
-def route_after_validation(state: Dict[str, Any]) -> str:
-    """
-    Route based on validation results.
-    """
-    needs_retry = state.get("needs_retry", False)
-    research_type = state.get("research_type", "general")
-
-    if needs_retry:
-        # Route back to appropriate research node for retry
-        return route_by_research_type(state)
-    else:
-        # Proceed to synthesis
-        return "synthesize_results"
-
-
-# ======================== Graph Construction ========================
-
-def create_live_research_workflow() -> StateGraph:
-    """
-    Create the live research workflow graph with AI and real tools.
-    """
-    # Initialize the graph with enhanced state schema
-    initial_state = {
-        "input": "",
-        "research_type": "",
-        "research_data": {},
-        "step_count": 0,
-        "retry_count": 0,
-        "validation_status": "",
-        "needs_retry": False,
-        "output": "",
-        "live_data_included": False,
-        "final_confidence": 0.0
-    }
-
-    graph = StateGraph(initial_state)
-
-    # Add all nodes
-    graph.add_node("analyze_query", analyze_query_node)
-    graph.add_node("weather_research", weather_research_node)
-    graph.add_node("financial_research", financial_research_node)
-    graph.add_node("news_research", news_research_node)
-    graph.add_node("technical_research", technical_research_node)
-    graph.add_node("general_research", general_research_node)
-    graph.add_node("validate_research", validate_research_node)
-    graph.add_node("synthesize_results", synthesize_results_node)
-    graph.add_node("error_handling", error_handling_node)
-
-    # Set entry point
-    graph.set_entry_point("analyze_query")
-
-    # Add conditional edges for research type routing
-    graph.add_conditional_edges(
-        "analyze_query",
-        route_by_research_type,
-        {
-            "weather_research": "weather_research",
-            "financial_research": "financial_research",
-            "news_research": "news_research",
-            "technical_research": "technical_research",
-            "general_research": "general_research"
-        }
-    )
-
-    # All research nodes go to validation
-    for research_node in ["weather_research", "financial_research", "news_research",
-                         "technical_research", "general_research"]:
-        graph.add_edge(research_node, "validate_research")
-
-    # Add conditional edges for validation routing
-    graph.add_conditional_edges(
-        "validate_research",
-        route_after_validation,
-        {
-            "weather_research": "weather_research",
-            "financial_research": "financial_research",
-            "news_research": "news_research",
-            "technical_research": "technical_research",
-            "general_research": "general_research",
-            "synthesize_results": "synthesize_results"
-        }
-    )
-
-    return graph
-
-
-# ======================== Demo Functions ========================
-
-async def run_single_demo(query: str, agent: GraphAgent):
-    """Run a single demo query and display results."""
-    print(f"\n{'='*60}")
-    print(f"🔍 Query: {query}")
-    print(f"{'='*60}")
-
-    try:
-        # Clear any previous state
-        agent.clear_state()
-
-        # Run the agent
-        result = await agent.run(query)
-
-        print(f"\n📊 Result:\n{result}")
-
-        # Show execution metadata
-        metadata = agent.get_execution_metadata()
-        if metadata.get("execution_successful"):
-            history = agent.get_execution_history()
-            print(f"\n🔄 Execution Path: {' → '.join([step['node'] for step in history])}")
-            print(f"⏱️  Total Steps: {len(history)}")
-
-            # Show live data integration info
-            final_state = getattr(agent, '_last_state', {})
-            if final_state.get("live_data_included"):
-                print(f"🌐 Live Data: Integrated with confidence {final_state.get('final_confidence', 0):.1%}")
-
-    except Exception as e:
-        print(f"\n❌ Error: {str(e)}")
-
-
-async def main():
-    """
-    Main demo function showcasing the live AI graph-based agent system.
-    """
-    print("🚀 Live AI Graph Agent Demo - Smart Research Assistant")
-    print("=" * 60)
-    print("🤖 Features: Live LLM Analysis + Real Web Search via Tavily")
-    print("=" * 60)
-
-    # Create the live research workflow graph
-    research_graph = create_live_research_workflow()
-
-    # Create a GraphAgent with the workflow and live ChatBot
-    try:
-        chatbot = create_configured_chatbot()
-        print("✅ ChatBot configured successfully")
-    except Exception as e:
-        print(f"⚠️  Warning: ChatBot configuration issue: {str(e)}")
-        print("   Demo will continue with fallback functionality")
-        chatbot = None
-
-    agent = GraphAgent(
-        name="live_research_assistant",
-        description="An intelligent research assistant using live AI and real web search tools",
-        system_prompt="You are a helpful research assistant that uses live AI analysis and real web search to provide comprehensive, up-to-date research results.",
-        llm=chatbot,
-        graph=research_graph,
-        preserve_state=False  # Start fresh for each query
-    )
-
-    # Demo queries showcasing live AI and real tool integration
-    demo_queries = [
-        "What's the current weather in Tokyo?",
-        "What are the latest Bitcoin price trends today?",
-        "What's the latest news about artificial intelligence developments?",
-        "How do I implement a binary search algorithm in Python?",
-        "Tell me about recent developments in quantum computing"
-    ]
-
-    print(f"\n🎯 Running {len(demo_queries)} demo queries with live AI and tools:")
-    print("   • LLM-powered query classification")
-    print("   • Real-time web search via tavily-search")
-    print("   • Live data validation and synthesis")
-
-    for query in demo_queries:
-        await run_single_demo(query, agent)
-        await asyncio.sleep(1)  # Brief pause between demos
-
-    print(f"\n{'='*60}")
-    print("✅ Live AI Graph Agent Demo Complete!")
-    print("\n🔍 Key Features Demonstrated:")
-    print("  • 🤖 LLM-powered query analysis and classification")
-    print("  • 🌐 Live web search integration via tavily-search tool")
-    print("  • 📊 Real-time data validation and quality control")
-    print("  • 🔄 Intelligent retry logic for failed searches")
-    print("  • 📝 Comprehensive synthesis of live research results")
-    print("  • ⚡ Error handling with graceful degradation")
-    print("  • 📈 Confidence scoring and metadata tracking")
-    print("\n🎉 The live AI graph-based agent system is working correctly!")
-    print("\n💡 Note: For full functionality, ensure:")
-    print("   • MCP server with tavily-search tool is running")
-    print("   • LLM API keys are properly configured")
-    print("   • Internet connection is available for live searches")
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
