+++ spoon_ai/graph/__init__.py
@@ -1,0 +1,41 @@
+"""
+spoon_ai.graph package
+
+Public facade for the graph engine. Import from here.
+"""
+# Re-export public API
+from .exceptions import (
+    GraphExecutionError,
+    NodeExecutionError,
+    StateValidationError,
+    CheckpointError,
+    GraphConfigurationError,
+    EdgeRoutingError,
+    InterruptError,
+)
+from .types import (
+    NodeContext,
+    NodeResult,
+    RouterResult,
+    ParallelBranchConfig,
+    Command,
+    StateSnapshot,
+)
+from .reducers import (
+    add_messages,
+    merge_dicts,
+    append_history,
+    union_sets,
+    validate_range,
+    validate_enum,
+)
+from .decorators import (
+    node_decorator,
+    router_decorator,
+)
+from .checkpointer import InMemoryCheckpointer
+
+# Engine and agent implementations (now within this package)
+from .engine import StateGraph, CompiledGraph, interrupt
+from .agent import GraphAgent, AgentStateCheckpoint, MockMemory
+
